refactor: simplify expo bootstrap
